--- WireDomain/Sources/WireDomain/Event Processing/ConversationEventProcessor/ConversationMessageAddEventProcessor/todo.txt
@@ -1,1 +1,0 @@
-move mls + proteus + shared files here
